Update week03.html
Populated style #0 (Stylesheet 1) and described changes. Added a fifth sheet to the template for even division of work. Moved the meta element with the charset attribute to the beginning of the document. Apparently the element containing the character encoding declaration must be serialized completely within the first 1024 bytes of the document or the validator fails.